aggiunti 8 test

+++ API/users/remove_user.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const validateToken = require("../validate_token");
 const User = require("../../models/users");
+const mongoose = require("mongoose");
 const router = express.Router();
 // Middleware per gestire i dati JSON nelle richieste
 router.use(express.json());
@@ -70,16 +71,22 @@ router.use(express.json());
  */
 router.delete("/remove_user/", validateToken, (req, res) => {
     let id = req.body.id;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        res.status(400).json({ message: "Invalid ID" });
+        return;
+    }
+
     User.findByIdAndDelete(id)
-      .then((data) => {
-        if(data == null) 
-            res.status(404).json({ message: "User Not Found" });
-        else
-            res.status(201).json({ message: "User Removed Successfully" });
-      })
-      .catch((error) => {
-        res.status(500).json({ message: error.message, type: "danger" });
-      });
-  });
+        .then((data) => {
+            if (data == null)
+                res.status(404).json({ message: "User Not Found" });
+            else
+                res.status(201).json({ message: "User Removed Successfully" });
+        })
+        .catch((error) => {
+            res.status(500).json({ message: error.message, type: "danger" });
+        });
+});
 
 module.exports = router;
